requirements.txt
# Core dependencies
agno>=0.3.0  # Agno framework for AI agents
google-genai>=0.1.0  # Google Gemini API
google-generativeai>=0.8.0  # Direct Google GenerativeAI client for streaming
fastapi>=0.104.0  # API framework
uvicorn[standard]>=0.24.0  # ASGI server
python-multipart>=0.0.6  # For form data handling
python-dotenv>=1.0.0  # Environment variable management

# Database
aiosqlite>=0.19.0  # Async SQLite for memory/storage

# API clients and integrations
requests>=2.31.0  # HTTP client
httpx>=0.25.0  # Async HTTP client for streaming API calls
PyGithub>=2.1.1  # GitHub API client
slack_sdk>=3.26.0  # Slack API client
duckduckgo-search>=4.0 # DuckDuckGo search tool dependency

# CLI Dependencies - Beautiful Terminal Interface
rich>=13.7.0  # Rich text and beautiful formatting in the terminal
typer>=0.9.0  # Modern CLI framework with auto-completion
click>=8.1.7  # Command line interface creation toolkit
prompt-toolkit>=3.0.43  # Beautiful interactive command-line interfaces
textual>=0.50.0  # Modern TUI framework for Python
colorama>=0.4.6  # Cross-platform colored terminal text
blessed>=1.20.0  # Easy, practical library for making terminal apps
inquirer>=3.1.3  # Collection of common interactive command line user interfaces

# Optional dependencies (uncomment if needed)
# mem0>=0.1.0  # Mem0 client for long-term memory (optional)
# pgvector>=0.2.4  # Vector storage for PostgreSQL (optional)
# psycopg>=3.1.12  # PostgreSQL driver (optional)

# Development dependencies
# pytest>=7.4.3  # Testing framework
# pytest-asyncio>=0.21.1  # Async testing support
# black>=23.11.0  # Code formatting
# isort>=5.12.0  # Import sorting 